pyproject.toml
[tool.poetry]
name = "fast_auth"
version = "0.1.0"
description = ""
authors = ["Douglas Braga <douglas.braga@modenaesilva.com.br>"]

[tool.poetry.dependencies]
python = "^3.9"
python-jose = "^3.3.0"
cryptography = "^35.0.0"
passlib = "^1.7.4"
bcrypt = "^3.2.0"
SQLAlchemy = "^1.4.26"
aiosqlite = "^0.17.0"
python-dotenv = "^0.19.1"
fastapi = "^0.70.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.6.0"
pytest-asyncio = "^0.16.0"
httpx = "^0.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
migrate = "fast_auth.migration:create_auth_tables_cli"
create_user = "fast_auth.auth.auth:create_user_cli"
